refactor ec2
