+++ stopinator/aws/utils.py
@@ -8,6 +8,7 @@ import calendar
 
 CONST_KEY_TIME_START="time:start"
 CONST_KEY_TIME_STOP="time:stop"
+CONST_KEY_ACTIVE_TILL = "stopinator:active:till"
 
 CONST_ASG_RESUME_KEY="stopinator:resume:time"
 CONST_STOPINATOR_STOP_TIME="stopinator:stop:time"
@@ -118,115 +119,6 @@ def get_time(arg,tags):
            r[1]=int(timepart[1])
 
    return r
-
-
-## stop instance
-def suspend_asg(name):
-    print "suspending asg:"+name
-    response = asgclient.suspend_processes(
-    	AutoScalingGroupName=name
-    )
-    print response
-
-
-def resume_asg(name):
-    print "Resuming asg :"+name
-    response = asgclient.resume_processes(
-    	AutoScalingGroupName=name,
-    )
-    print response
-    print "remove schedule"
-    response = asgclient.delete_tags(
-    	Tags=[{
-            "ResourceId": name,
-            "ResourceType":"auto-scaling-group",
-            "Key": CONST_ASG_RESUME_KEY,
-        }]
-    )
-    print response
-
-def generate_asg_instance(tz):
-    asgList={"1":"name"}
-
-    print "creating autoscaling instance map"
-
-    response = asgclient.describe_auto_scaling_groups()
-
-    #print response
-    #nextToken = response['NextToken']
-    asgs = response['AutoScalingGroups']
-    for asg in asgs:
-        name = asg['AutoScalingGroupName']
-        tags = asg["Tags"]
-        ## starting suspended asgs based on tiem
-        v = get_time(CONST_ASG_RESUME_KEY,tags)
-        if not v:
-           print "no asg schedule(nothing to resume)"
-        else:
-           c = current_time(tz)
-           if c[0] > v[0]:
-              resume_asg(name)
-           if c[0]==v[0] and c[1] >= v[1]:
-              resume_asg(name)
-        # end asg stuff
-
-        #print asg['AutoScalingGroupName'],'\n'
-        for instance in asg['Instances']:
-            iid= instance['InstanceId']
-            asgList[iid] = name
-    return asgList
-
-def stop_instance(instance,asglist,tz):
-    iid = instance.get("InstanceId")
-    current = current_time(tz)
-    ec2.create_tags(Resources=[iid], Tags=[{"Key":"stopinator:stop:time","Value":current[2]}])
-    if iid in asglist:
-        print "has associated asg.need to suspend it first"
-        asg = asglist[iid]
-        suspend_asg(asg)
-    print ec2.stop_instances(InstanceIds=[iid])
-
-
-
-##start instanc
-def start_instance(instance,asglist,tz):
-    iid = instance.get("InstanceId")
-    print ec2.start_instances(InstanceIds=[iid])
-    ctime = current_time(tz)
-    cdate = ctime[2].split("T")
-
-
-    current = get_time("time:start",instance.get("Tags"))
-    hh=""
-    if current[0] < 10:
-       hh="0"+`current[0]`
-    else:
-       hh=""+`current[0]`
-    cdatetime = cdate[0]+"T"+hh+":"
-
-    if current[1] < 10:
-       cdatetime = cdatetime + "0"+`current[1]`
-    else:
-       cdatetime = cdatetime+`current[1]`
-
-    ec2.create_tags(Resources=[iid], Tags=[{"Key":CONST_STOPINATOR_START_TIME,"Value":cdatetime}])
-    #schedule asg resume if instance is part of ASG (need to make sure it resume after all the instance are on)
-    if iid in asglist:
-       name = asglist[iid];
-       ctime[1] = ctime[1]+4
-       lt = ""+`ctime[0]`+":"+`ctime[1]`
-       print "asg schedule time:"+lt
-       response = asgclient.create_or_update_tags(
-    		Tags=[{
-            	    "ResourceId": name,
-                    "Key": CONST_ASG_RESUME_KEY,
-                    "Value":lt,
-                    "ResourceType":"auto-scaling-group",
-                    "PropagateAtLaunch": False
-               }]
-       )
-       print response
-
 
 def can_start(current, tags):
   ch = current[0]
