+++ stopinator/lambda_function.py
@@ -5,15 +5,15 @@ import boto3
 import datetime
 import time
 import calendar
-from aws import utils
+from aws import utils,ec2
 
 
 
 
 CONST_TZ = "Australia/NSW"
 
-asgclient = boto3.client('autoscaling',region_name="ap-southeast-2")
-ec2 = boto3.client('ec2',region_name='ap-southeast-2')
+#asgclient = boto3.client('autoscaling',region_name="ap-southeast-2")
+#ec2 = boto3.client('ec2',region_name='ap-southeast-2')
 
 
 
@@ -27,7 +27,7 @@ def lambda_handler(event, context):
     tz = CONST_TZ
     print("event content:",event)
     pattern = ["122acn.*"]
-    asg_instance = utils.generate_asg_instance(tz)
+    asg_instance = ec2.generate_asg_instance(tz)
     if not event:
         print "Loading default"
     else:
@@ -44,9 +44,9 @@ def lambda_handler(event, context):
 
     current = utils.current_time(tz)
 
-    filters = utils.instance_filter(pattern)
-    print "Using filters",filters
-    reservations=ec2.describe_instances(Filters=filters)
+    #filters = utils.instance_filter(pattern)
+    #print "Using filters",filters
+    reservations=ec2.list_instances(Matcher=pattern)
     print "found  "+`len(reservations['Reservations'])`+" Reservation matches"
 
     for r in reservations['Reservations']:
@@ -75,7 +75,8 @@ def lambda_handler(event, context):
             if stateId == 16:
                if utils.can_stop(current,tags):
                   print "STOPING INSTANCE",iid
-                  utils.stop_instance(i,asg_instance,tz)
+                  #update_stopinator_status(iid,tz,tags)
+                  ec2.stop_instance(i,asg_instance,tz)
                   executeStop = True
             #start condition
             if not executeStop:
@@ -83,6 +84,6 @@ def lambda_handler(event, context):
                   print "Instance stateId:"+`stateId`
                   if utils.can_start(current,tags):
                      print "STARTING INSTANCE:"+iid
-                     utils.start_instance(i,asg_instance,tz)
+                     ec2.start_instance(i,asg_instance,tz)
 
     return "OK"
